Refactor type handling and improve cast semantics
- Updated LLVM bindings to use `size_t` for length parameters in `LLVMGetValueName2`.
- Introduced `widen_integer_init` to handle integer literal initializers with explicit casts.
- Enhanced `DeclarationEmitter` to correctly handle constant folding for cast expressions and scope resolutions.
- Added checks for function type compatibility in `CallResolver` for null literals.
- Improved variable declaration handling in `sema` to ensure global constants are resolved correctly.
- Added utility functions in `TypeUtils` to check for function types and improve type compatibility checks.
- Enhanced `TypeChecker` to recognize closures as compatible with function types.
- Added new tests for explicit casting behavior and module-level constants initialized from enum variants.
- Introduced negative tests for assignment mismatches involving function pointers and integer types.

+++ compiler/llvm_bindings.h
@@ -196,7 +196,10 @@ LLVMTypeRef LLVMVectorType(LLVMTypeRef ElementType, unsigned ElementCount);
  * =================================================================== */
 
 LLVMTypeRef    LLVMTypeOf(LLVMValueRef Val);
-const char    *LLVMGetValueName2(LLVMValueRef Val, unsigned long *Length);
+/* `size_t *`, not `unsigned long *`: LLVM writes a size_t through this pointer,
+ * and `unsigned long` is 32-bit on Windows (LLP64), so the binding would claim
+ * a 4-byte slot for an 8-byte store.  The caller already passes a `u64`. */
+const char    *LLVMGetValueName2(LLVMValueRef Val, size_t *Length);
 void           LLVMSetValueName2(LLVMValueRef Val, const char *Name, unsigned long NameLen);
 void           LLVMDumpValue(LLVMValueRef Val);
 char          *LLVMPrintValueToString(LLVMValueRef Val);
